add Quote Everyday! python script.
script logs in on twitter using consumer key and secret and access token
and -secret. Point the <.login> path/file to the path / filename you use.

place the tweet you want to quote in the kwoots.txt file, the python
script reads it from there (if you change the path of the file, change it
in the python script accordingly.

I made a cron on my raspi 3 which execute the file daily at 15:00 local

Feel free to at quotes for daily quotes by forking, updating and PR
to kwoots.txt!

NOTE: made on python v 3.6!

Signed-off-by: Erik Castricum <git@cas-online.nl>