track tweet deletion
in this example we track the tweet deletion behaviour
of a fake Dutch rapper, appears to be living in Spain.

He likes himself so much, that he often retweets his own
tweet instantly.

Note: last minute added error handler, the script was
probably crashing on a query in json to user, upon deleting
a tweet.

Signed-off-by: Erik Castricum <git@cas-online.nl>